refactor: #1193 PremiumModal/MonthlyRewardModal → Dialog リネーム (後方互換 stub 維持)
#1175 UI 監査で指摘された「Dialog primitive を内部利用しているが Modal 命名」
の揺れを解消。

## 変更

- PremiumModal.svelte → PremiumDialog.svelte (Dialog primitive 利用の実体)
- MonthlyRewardModal.svelte → MonthlyRewardDialog.svelte
- 旧パス (PremiumModal.svelte / MonthlyRewardModal.svelte) に薄い
  re-export stub を配置 (ADR-0001 後方互換 = 1 サイクル維持)
- 呼び出し側 2 箇所を更新:
  - src/lib/ui/components/PremiumBadge.svelte (dynamic import)
  - src/routes/(child)/[uiMode=uiMode]/home/+page.svelte (static import)

## Test plan

- npx biome check: clean
- npx svelte-check: 0 errors (pre-existing 45 warnings unchanged)
- npx vitest run: 3630/3630 passed

## Closes

Closes #1193
